requirements.txt


pythonnet>=3.0.0

# FastAPI 相关依赖
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

influxdb-client>=1.36.0


sqlalchemy>=2.0.0
alembic>=1.12.0

bcrypt>=4.0.0


PyJWT>=2.8.0


python-dotenv>=1.0.0


loguru>=0.7.0


pydantic>=2.0.0


python-dateutil>=2.8.0


apscheduler>=3.10.0

